Update Dreamstore landing to StoreFramework and adjust tachyons tokens
