+++ styles/configs/style.json
@@ -200,7 +200,7 @@
       "muted-5": "#f2f4f5"
     },
     "on": {
-      "base": "#142032",
+      "base": "#ffffff",
       "base--inverted": "#ffffff",
       "action-primary": "#ffffff",
       "action-secondary": "#F4F3F8",
